requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Retrievers / index build (Phase 0)
rank-bm25>=0.2.2
duckdb>=0.9.0
sentence-transformers>=2.2.0

# LLM (pipeline answer generation)
openai>=1.0.0

# Router / ML (Phase 2 learned router)
torch>=2.0
transformers>=4.30.0
numpy>=1.24.0

# API (optional)
# fastapi
# uvicorn
